Create random-maze-using-ASCII.c
We define a constant SIZE to specify the dimensions of the maze. Adjusting this value will change the size of the maze.

The generateMaze function creates a maze with walls (#) and empty spaces ( ). It also randomly selects a start point (S) and an end point (E) for the maze.

The printMaze function displays the maze on the console.

In the main function, we call generateMaze to create the maze and then printMaze to display it.


#haktoberfest23